news_agent/requirements.txt
feedparser==6.0.10
ollama==0.1.7
apscheduler==3.10.4
pywhatkit==5.4
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8

